Move `PrunableTree` and `RetentionFlags` into `prunable` submodule
